GCA-29 Migrate DAO to Hibernate - refactoring
