+++ src/test/java/com/gca/dao/TrainerDAOTest.java
@@ -1,135 +1,75 @@
 package com.gca.dao;
 
 import com.gca.dao.impl.TrainerDAOImpl;
-import com.gca.exception.DaoException;
 import com.gca.model.Trainer;
 import com.gca.model.TrainingType;
 import com.gca.model.User;
+import com.github.database.rider.core.api.dataset.DataSet;
+import org.hibernate.Session;
 import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
+import org.springframework.test.context.ContextConfiguration;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
-class TrainerDAOTest extends AbstractDAOTest {
-
-    private static final Long USER_ID = 1001L;
-    private static final Long TYPE_ID = 991L;
-
-    private static final String USERNAME = "traineruser";
-    private static final String FIRSTNAME = "Jane";
-    private static final String LASTNAME = "Smith";
-    private static final String PASSWORD = "trainerpass";
-    private static final String TYPE_NAME = "Fitness";
-    private static final boolean IS_ACTIVE = true;
-
-    @InjectMocks
-    private TrainerDAOImpl dao;
+@ContextConfiguration(classes = TrainerDAOImpl.class)
+@DataSet(value = "dataset/trainer/trainer-data.xml", cleanBefore = true, cleanAfter = true, transactional = true)
+public class TrainerDAOTest extends BaseIntegrationTest<TrainerDAOImpl> {
 
     @Test
-    void shouldSuccessfullyCreateTrainer() {
-        Trainer expected = buildTrainer(1L, USER_ID, USERNAME);
+    void shouldSuccessfullyFindTrainer() {
+        Trainer expected = sessionFactory.getCurrentSession().find(Trainer.class, 1L);
 
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
+        Trainer actual = dao.getById(expected.getId());
 
-        Trainer actual = dao.create(expected);
-
-        verify(sessionFactory).getCurrentSession();
-        verify(session).persist(expected);
-
-        assertEquals(expected, actual);
-        assertEquals(expected.getId(), actual.getId());
-        assertEquals(expected.getSpecialization(), actual.getSpecialization());
-        assertEquals(expected.getUser().getId(), actual.getUser().getId());
+        assertNotNull(actual, "Loaded trainer should not be null");
+        assertEquals(expected.getSpecialization().getId(), actual.getSpecialization().getId());
+        assertEquals(expected.getUser().getUsername(), actual.getUser().getUsername());
+        assertEquals(expected.getUser().getFirstName(), actual.getUser().getFirstName());
+        assertEquals(expected.getUser().getLastName(), actual.getUser().getLastName());
     }
 
     @Test
-    void shouldReturnTrainerById() {
-        Long trainerId = 10L;
-        Trainer expected = buildTrainer(trainerId, 20L, "user10");
+    void shouldNotFindTrainer() {
+        Trainer found = dao.getById(99L);
+        assertNull(found, "Trainer should be null");
+    }
 
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
-        when(session.find(Trainer.class, trainerId)).thenReturn(expected);
+    @Test
+    @DataSet(value = "dataset/trainer/trainer-creation-data.xml", cleanBefore = true, cleanAfter = true, transactional = true)
+    void shouldCreateTrainer() {
+        Trainer expected = buildTrainerFromExistingUserAndSpecialization();
 
-        Trainer actual = dao.getById(trainerId);
+        Trainer actual = dao.create(expected);
 
-        verify(sessionFactory).getCurrentSession();
-        verify(session).find(Trainer.class, trainerId);
+        assertNotNull(actual.getId(), "Created trainer should have ID");
 
-        assertEquals(expected, actual);
-        assertEquals(expected.getId(), actual.getId());
-        assertEquals(expected.getSpecialization(), actual.getSpecialization());
-        assertEquals(expected.getUser().getId(), actual.getUser().getId());
+        assertEquals(expected.getUser().getUsername(), actual.getUser().getUsername());
+        assertEquals(expected.getSpecialization().getId(), actual.getSpecialization().getId());
+        assertEquals(expected.getUser().getFirstName(), actual.getUser().getFirstName());
+        assertEquals(expected.getUser().getLastName(), actual.getUser().getLastName());
     }
 
     @Test
     void shouldUpdateTrainer() {
-        Long trainerId = 5L;
-        Trainer trainer = buildTrainer(trainerId, 105L, "upduser");
-
-        TrainingType newType = TrainingType.builder()
-                .id(123L)
-                .name("Yoga")
-                .build();
-
-        Trainer expected = trainer.toBuilder()
-                .specialization(newType)
-                .build();
-
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
-        when(session.find(Trainer.class, trainerId)).thenReturn(trainer);
-        when(session.merge(expected)).thenReturn(expected);
+        Trainer expected = sessionFactory.getCurrentSession().find(Trainer.class, 1L);
+        expected.getUser().setFirstName("UpdatedName");
 
         Trainer actual = dao.update(expected);
 
-        verify(sessionFactory).getCurrentSession();
-        verify(session).find(Trainer.class, trainerId);
-        verify(session).merge(expected);
-
-        assertEquals(expected, actual);
-        assertEquals(expected.getId(), actual.getId());
-        assertEquals(expected.getSpecialization(), actual.getSpecialization());
-        assertEquals(expected.getUser().getId(), actual.getUser().getId());
+        assertEquals(expected.getUser().getFirstName(), actual.getUser().getFirstName());
+        assertEquals(expected.getSpecialization().getId(), actual.getSpecialization().getId());
+        assertEquals(expected.getUser().getLastName(), actual.getUser().getLastName());
     }
 
-    @Test
-    void shouldThrowWhenUpdatingNonExistingTrainer() {
-        Long trainerId = 5L;
-        Trainer trainer = buildTrainer(trainerId, 105L, "upduser");
-
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
-        when(session.find(Trainer.class, trainerId)).thenReturn(null);
-
-        assertThrows(DaoException.class, () -> dao.update(trainer));
-
-        verify(sessionFactory).getCurrentSession();
-        verify(session).find(Trainer.class, trainerId);
-        verify(session, never()).merge(any());
-    }
-
-    private Trainer buildTrainer(Long trainerId, Long userId, String username) {
+    private Trainer buildTrainerFromExistingUserAndSpecialization() {
+        Session session = sessionFactory.getCurrentSession();
+        User existingUser = session.find(User.class, 2L);
+        TrainingType specialization = session.find(TrainingType.class, 1L);
         return Trainer.builder()
-                .id(trainerId)
-                .user(buildUser(userId, username))
-                .specialization(TrainingType.builder()
-                        .id(TYPE_ID)
-                        .name(TYPE_NAME)
-                        .build())
-                .build();
-    }
-
-    private User buildUser(Long userId, String username) {
-        return User.builder()
-                .id(userId)
-                .username(username)
-                .firstName(FIRSTNAME)
-                .lastName(LASTNAME)
-                .password(PASSWORD)
-                .isActive(IS_ACTIVE)
+                .user(existingUser)
+                .specialization(specialization)
                 .build();
     }
 }
