+++ src/test/java/com/gca/dao/UserDAOTest.java
@@ -3,174 +3,124 @@ package com.gca.dao;
 import com.gca.dao.impl.UserDAOImpl;
 import com.gca.exception.DaoException;
 import com.gca.model.User;
+import com.github.database.rider.core.api.dataset.DataSet;
 import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
+import org.springframework.test.context.ContextConfiguration;
 
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class UserDAOTest extends AbstractDAOTest {
-
-    private static final Long USER_ID = 1001L;
-    private static final String USERNAME = "testuser";
-    private static final String FIRSTNAME = "John";
-    private static final String LASTNAME = "Doe";
-    private static final String PASSWORD = "password";
-    private static final boolean IS_ACTIVE = true;
-
-    @InjectMocks
-    private UserDAOImpl dao;
+@ContextConfiguration(classes = UserDAOImpl.class)
+@DataSet(value = "dataset/user/user-data.xml", cleanBefore = true, cleanAfter = true, transactional = true)
+public class UserDAOTest extends BaseIntegrationTest<UserDAOImpl> {
+    private static final Long EXISTING_ID = 1L;
+    private static final String EXISTING_USERNAME = "john.doe";
 
     @Test
-    void shouldSuccessfullyCreateUser() {
-        User expected = buildUser();
+    void shouldFindUserById() {
+        User expected = sessionFactory.getCurrentSession().find(User.class, EXISTING_ID);
 
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
+        User actual = dao.getById(EXISTING_ID);
 
-        User actual = dao.create(expected);
-
-        verify(sessionFactory).getCurrentSession();
-        verify(session).persist(expected);
-
-        assertEquals(expected, actual);
+        assertNotNull(actual);
+        assertEquals(expected.getUsername(), actual.getUsername());
+        assertEquals(expected.getPassword(), actual.getPassword());
+        assertEquals(expected.getFirstName(), actual.getFirstName());
     }
 
     @Test
-    void shouldReturnUserById() {
-        User expected = buildUser();
-
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
-        when(session.find(User.class, USER_ID)).thenReturn(expected);
-
-        User actual = dao.getById(USER_ID);
-
-        verify(sessionFactory).getCurrentSession();
-        verify(session).find(User.class, USER_ID);
-
-        assertEquals(expected, actual);
+    void shouldNotFindUserById() {
+        User actual = dao.getById(999L);
+        assertNull(actual);
     }
 
     @Test
-    void shouldUpdateUser() {
-        User existing = buildUser();
-        User expected = existing.toBuilder()
-                .firstName("Updated")
-                .build();
+    void shouldFindUserByUsername() {
+        User actual = dao.getByUsername(EXISTING_USERNAME);
 
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
-        when(session.find(User.class, USER_ID)).thenReturn(existing);
-        when(session.merge(expected)).thenReturn(expected);
-
-        User actual = dao.update(expected);
-
-        verify(sessionFactory).getCurrentSession();
-        verify(session).find(User.class, USER_ID);
-        verify(session).merge(expected);
-
-        assertEquals(expected, actual);
+        assertNotNull(actual);
+        assertEquals(EXISTING_USERNAME, actual.getUsername());
     }
 
     @Test
-    void shouldThrowWhenUpdatingNonExistingUser() {
-        User nonExisting = buildUser();
-
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
-        when(session.find(User.class, USER_ID)).thenReturn(null);
-
-        assertThrows(DaoException.class, () -> dao.update(nonExisting));
-
-        verify(sessionFactory).getCurrentSession();
-        verify(session).find(User.class, USER_ID);
-        verify(session, never()).merge(any());
-    }
-
-    @Test
-    void shouldDeleteUser() {
-        User existing = buildUser();
-
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
-        when(session.find(User.class, USER_ID)).thenReturn(existing);
-
-        dao.delete(USER_ID);
-
-        verify(sessionFactory).getCurrentSession();
-        verify(session).find(User.class, USER_ID);
-        verify(session).remove(existing);
-    }
-
-    @Test
-    void shouldThrowWhenDeletingNonExistingUser() {
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
-        when(session.find(User.class, USER_ID)).thenReturn(null);
-
-        assertThrows(DaoException.class, () -> dao.delete(USER_ID));
-
-        verify(sessionFactory).getCurrentSession();
-        verify(session).find(User.class, USER_ID);
-        verify(session, never()).remove(any());
-    }
-
-    @Test
-    void shouldReturnUserByUsername() {
-        User expected = buildUser();
-
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
-
-        var query = mock(org.hibernate.query.Query.class);
-
-        when(session.createQuery(
-                "FROM User u WHERE u.username = :username", User.class))
-                .thenReturn(query);
-
-        when(query.setParameter("username", USERNAME)).thenReturn(query);
-        when(query.uniqueResult()).thenReturn(expected);
-
-        User actual = dao.getByUsername(USERNAME);
-
-        verify(sessionFactory).getCurrentSession();
-        verify(session).createQuery(anyString(), eq(User.class));
-        verify(query).setParameter("username", USERNAME);
-        verify(query).uniqueResult();
-
-        assertEquals(expected, actual);
+    void shouldReturnNullWhenUserNotFoundByUsername() {
+        User actual = dao.getByUsername("unknown.user");
+        assertNull(actual);
     }
 
     @Test
     void shouldReturnAllUsernames() {
-        List<String> usernames = List.of("user1", "user2", "user3");
+        List<String> usernames = dao.getAllUsernames();
 
-        var query = mock(org.hibernate.query.Query.class);
-
-        when(sessionFactory.getCurrentSession()).thenReturn(session);
-        when(session.createQuery(anyString(), eq(String.class))).thenReturn(query);
-        when(query.getResultList()).thenReturn(usernames);
-
-        List<String> actual = dao.getAllUsernames();
-
-        verify(sessionFactory).getCurrentSession();
-        verify(session).createQuery(anyString(), eq(String.class));
-        verify(query).getResultList();
-
-        assertEquals(usernames, actual);
+        assertNotNull(usernames);
+        assertTrue(usernames.contains("john.doe"));
+        assertTrue(usernames.contains("ghost.user"));
+        assertEquals(2, usernames.size());
     }
 
-    private User buildUser() {
-        return User.builder()
-                .id(UserDAOTest.USER_ID)
-                .username(UserDAOTest.USERNAME)
-                .firstName(FIRSTNAME)
-                .lastName(LASTNAME)
-                .password(PASSWORD)
-                .isActive(IS_ACTIVE)
+    @Test
+    @DataSet(value = "dataset/user/no-user-data.xml", cleanBefore = true, cleanAfter = true, transactional = true)
+    void shouldCreateUser() {
+        User newUser = User.builder()
+                .firstName("New")
+                .lastName("User")
+                .username("new.user")
+                .password("password")
+                .isActive(true)
                 .build();
+
+        User actual = dao.create(newUser);
+
+        assertNotNull(actual.getId(), "Created user should have ID");
+        assertEquals(newUser.getUsername(), actual.getUsername());
+        assertEquals(newUser.getPassword(), actual.getPassword());
+        assertEquals(newUser.getIsActive(), actual.getIsActive());
+    }
+
+    @Test
+    void shouldUpdateUser() {
+        User user = sessionFactory.getCurrentSession().find(User.class, EXISTING_ID);
+        user.setFirstName("UpdatedName");
+
+        User updated = dao.update(user);
+
+        assertEquals("UpdatedName", updated.getFirstName());
+    }
+
+    @Test
+    void shouldThrowWhenUpdatingNonExistingUser() {
+        User nonExisting = User.builder()
+                .id(999L)
+                .firstName("Ghost")
+                .lastName("User")
+                .username("ghost.nonexistent")
+                .password("pass")
+                .isActive(true)
+                .build();
+
+        assertThrows(DaoException.class, () -> dao.update(nonExisting));
+    }
+
+    @Test
+    void shouldDeleteUserById() {
+        User existingUser = sessionFactory.getCurrentSession().find(User.class, EXISTING_ID);
+        assertNotNull(existingUser);
+
+        dao.delete(existingUser.getId());
+
+        boolean isDeleted = sessionFactory.getCurrentSession()
+                .find(User.class, EXISTING_ID) == null;
+
+        assertTrue(isDeleted, "User should be deleted");
+    }
+
+    @Test
+    void shouldThrowWhenDeletingNonExistingUser() {
+        assertThrows(DaoException.class, () -> dao.delete(999L));
     }
 }
