--- src/test/java/com/gca/dao/integration/TrainerDAOIT.java
@@ -1,81 +1,0 @@
-package com.gca.dao.integration;
-
-import com.gca.dao.impl.TrainerDAOImpl;
-import com.gca.model.Trainer;
-import com.gca.model.TrainingType;
-import com.gca.model.User;
-import com.github.database.rider.core.api.dataset.DataSet;
-import org.hibernate.Session;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
-
-@ContextConfiguration(classes = TrainerDAOImpl.class)
-@DataSet(value = "dataset/trainer/trainer-data.xml", cleanBefore = true, cleanAfter = true, transactional = true)
-public class TrainerDAOIT extends AbstractDAOIT<TrainerDAOImpl> {
-
-    @Autowired
-    void setDao(TrainerDAOImpl dao) {
-        this.dao = dao;
-    }
-
-    @Test
-    void shouldSuccessfullyFindTrainer() {
-        Trainer expected = sessionFactory.getCurrentSession().find(Trainer.class, 1L);
-
-        Trainer actual = dao.getById(expected.getId());
-
-        assertNotNull(actual, "Loaded trainer should not be null");
-        assertEquals(expected.getSpecialization().getId(), actual.getSpecialization().getId());
-        assertEquals(expected.getUser().getUsername(), actual.getUser().getUsername());
-        assertEquals(expected.getUser().getFirstName(), actual.getUser().getFirstName());
-        assertEquals(expected.getUser().getLastName(), actual.getUser().getLastName());
-    }
-
-    @Test
-    void shouldNotFindTrainer() {
-        Trainer found = dao.getById(99L);
-        assertNull(found, "Trainer should be null");
-    }
-
-    @Test
-    @DataSet(value = "dataset/trainer/trainer-creation-data.xml", cleanBefore = true, cleanAfter = true, transactional = true)
-    void shouldCreateTrainer() {
-        Trainer expected = buildTrainerFromExistingUserAndSpecialization();
-
-        Trainer actual = dao.create(expected);
-
-        assertNotNull(actual.getId(), "Created trainer should have ID");
-
-        assertEquals(expected.getUser().getUsername(), actual.getUser().getUsername());
-        assertEquals(expected.getSpecialization().getId(), actual.getSpecialization().getId());
-        assertEquals(expected.getUser().getFirstName(), actual.getUser().getFirstName());
-        assertEquals(expected.getUser().getLastName(), actual.getUser().getLastName());
-    }
-
-    @Test
-    void shouldUpdateTrainer() {
-        Trainer expected = sessionFactory.getCurrentSession().find(Trainer.class, 1L);
-        expected.getUser().setFirstName("UpdatedName");
-
-        Trainer actual = dao.update(expected);
-
-        assertEquals(expected.getUser().getFirstName(), actual.getUser().getFirstName());
-        assertEquals(expected.getSpecialization().getId(), actual.getSpecialization().getId());
-        assertEquals(expected.getUser().getLastName(), actual.getUser().getLastName());
-    }
-
-    private Trainer buildTrainerFromExistingUserAndSpecialization() {
-        Session session = sessionFactory.getCurrentSession();
-        User existingUser = session.find(User.class, 2L);
-        TrainingType specialization = session.find(TrainingType.class, 1L);
-        return Trainer.builder()
-                .user(existingUser)
-                .specialization(specialization)
-                .build();
-    }
-}
