--- src/test/resources/db/changelog/init/test-changeset.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<databaseChangeLog
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-        xsi:schemaLocation="
-        http://www.liquibase.org/xml/ns/dbchangelog
-        http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.18.xsd">
-
-    <changeSet id="GCA-29-test-init-ddl" author="ArturGoz">
-        <sqlFile path="test-db-ddl.sql"
-                 relativeToChangelogFile="true"
-                 encoding="UTF-8"/>
-    </changeSet>
-</databaseChangeLog>
